requirements.txt
# Core Django
Django==5.2.7
asgiref==3.9.2

# Database
psycopg2-binary==2.9.11
dj-database-url==2.1.0

# Static Files
whitenoise==6.6.0

# Excel Export
openpyxl==3.1.5
et-xmlfile==2.0.0

# Image Processing
Pillow==12.0.0
opencv-python-headless==4.9.0.80
numpy==1.26.4

# CORS Headers (for Android app API)
django-cors-headers==4.3.1

# Tailwind CSS Integration
django-tailwind[reload]==3.8.0

# Environment Variables
python-decouple==3.8

# WSGI Server
gunicorn==21.2.0

# Utilities
sqlparse==0.5.3
tzdata==2025.2
python-dateutil==2.8.2
pytz==2023.3

# AI Meter Reading (Claude Vision API)
anthropic==0.40.0
